refactor: Modularize armory screen logic and move to dedicated file
